Gate HTTP/TLS behind an `http` feature, confined to services::http
Collect every ureq/TLS use in the editor into a single new module,
`services::http`, with a real implementation gated on the new `http`
feature and a no-op/error stub when it is off. The three callers (update
checker, telemetry, and the editor.httpFetch plugin API) now call plain
functions with no cfg at the call sites.

Building without `http` (part of the default set) drops the entire TLS
stack — ureq + rustls + ring + webpki — from the binary. Combined with the
existing minimal feature set this is:

    cargo build --profile min-size --no-default-features --features runtime

Also silence the now-legitimately-unused highlight-capture tables in
fresh-languages when zero tree-sitter grammars are enabled. The default
release build is unchanged (default now includes `http`).
